requirements.txt
langgraph 
langsmith 
langchain_groq 
langchain_community 
langchain 
langchain-pinecone 
pinecone-client 
arxiv 
Wikipedia
arxiv
requests
PyMuPDF
nltk
torch 
torchvision 
torchaudio
sentence-transformers
transformers 
accelerate 
bitsandbytes
numpy
pandas